// types/tracking/meta/MetaContentItem.ts
// Path: types/tracking/meta/MetaContentItem.ts

export type MetaContentItem = {
  id: string
  quantity: number
  item_price?: number | undefined
  item_name?: string | undefined
  item_brand?: string | undefined
  item_category?: string | undefined
  item_variant?: string | undefined
  item_list_id?: string | undefined
  item_list_name?: string | undefined
  index?: number | undefined
  title?: string | undefined
  content_name?: string | undefined
  content_category?: string | undefined
}